RAGWITHGEMINI/RAG_LANGCHAIN_GEMINIAPI_FAISS.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Get the API key from environment variables
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')

if not GEMINI_API_KEY:
    raise ValueError("Please set the GEMINI_API_KEY environment variable in your .env file")
