RAGWITHGEMINI/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
faiss-cpu==1.7.4 